Add path echo to browser capability
The server can now echo a request to both a client running in the
terminal and to a browser. This was accomplished by analyzing the
incoming request for the string "GET". If "GET" is present in the
request the server assumes the request is coming from the browser and
will send the necessary status line as the response along with the path.

As might already be glaringly obvious, the server can easily be fooled
into thinking a request from the terminal is actually a browser request
if the first request from the terminal contains "GET". This will then
break the server.